Add more tests for Connect handlers
