+++ packages/connect-node-test/src/crosstest/custom_metadata.spec.ts
@@ -13,21 +13,23 @@
 // limitations under the License.
 
 import {
-  decodeBinaryHeader,
-  encodeBinaryHeader,
   createCallbackClient,
   createPromiseClient,
+  decodeBinaryHeader,
+  encodeBinaryHeader,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import {
   SimpleRequest,
   SimpleResponse,
 } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("custom_metadata", function () {
-  describeTransports(crosstestTransports, (transport) => {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransports((transport) => {
     const size = 314159;
     const binaryValue = new Uint8Array([0xab, 0xab, 0xab]);
     const requestHeaders = {
@@ -58,7 +60,7 @@ describe("custom_metadata", function () {
       }
     }
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
       let responseHeaders: Headers | undefined;
       let responseTrailers: Headers | undefined;
       const response = await client.unaryCall(request, {
@@ -75,7 +77,7 @@ describe("custom_metadata", function () {
       expectResponseTrailers(responseTrailers);
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
       let responseHeaders: Headers | undefined;
       let responseTrailers: Headers | undefined;
       client.unaryCall(
@@ -99,4 +101,6 @@ describe("custom_metadata", function () {
       );
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
