+++ packages/connect-node-test/src/crosstest/custom_metadata_server_streaming.spec.ts
@@ -13,21 +13,23 @@
 // limitations under the License.
 
 import {
-  decodeBinaryHeader,
-  encodeBinaryHeader,
   createCallbackClient,
   createPromiseClient,
+  decodeBinaryHeader,
+  encodeBinaryHeader,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import {
   StreamingOutputCallRequest,
   StreamingOutputCallResponse,
 } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("custom_metadata_server_streaming", function () {
-  describeTransports(crosstestTransports, (transport) => {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransports((transport) => {
     const size = 314159;
     const binaryValue = new Uint8Array([0xab, 0xab, 0xab]);
     const requestHeaders = {
@@ -55,7 +57,7 @@ describe("custom_metadata_server_streaming", function () {
       }
     }
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
       let responseHeaders: Headers | undefined;
       let responseTrailers: Headers | undefined;
       for await (const response of client.streamingOutputCall(request, {
@@ -73,7 +75,7 @@ describe("custom_metadata_server_streaming", function () {
       expectResponseTrailers(responseTrailers);
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
       let responseHeaders: Headers | undefined;
       let responseTrailers: Headers | undefined;
       client.streamingOutputCall(
@@ -99,4 +101,6 @@ describe("custom_metadata_server_streaming", function () {
       );
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
