+++ packages/connect-node-test/src/crosstest/fail_server_streaming.spec.ts
@@ -13,23 +13,25 @@
 // limitations under the License.
 
 import {
+  Code,
   ConnectError,
+  connectErrorDetails,
   createCallbackClient,
   createPromiseClient,
-  connectErrorDetails,
-  Code,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import {
   ErrorDetail,
   StreamingOutputCallRequest,
   StreamingOutputCallResponse,
 } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("fail_server_streaming", () => {
-  function expectError(err: unknown, transportName: string) {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  function expectError(err: unknown) {
     const expectedErrorDetail = new ErrorDetail({
       reason: "soirée 🎉",
       domain: "connect-crosstest",
@@ -38,14 +40,11 @@ describe("fail_server_streaming", () => {
     if (err instanceof ConnectError) {
       expect(err.code).toEqual(Code.ResourceExhausted);
       expect(err.rawMessage).toEqual("soirée 🎉");
-      // the experimental gRPC transport does not implement error details
-      if (transportName !== "gRPC transport") {
-        const details = connectErrorDetails(err, ErrorDetail);
-        expect(details.length).toEqual(1);
-        expect(details[0]).toBeInstanceOf(ErrorDetail);
-        if (details[0] instanceof ErrorDetail) {
-          expect(expectedErrorDetail.equals(details[0])).toBeTrue();
-        }
+      const details = connectErrorDetails(err, ErrorDetail);
+      expect(details.length).toEqual(1);
+      expect(details[0]).toBeInstanceOf(ErrorDetail);
+      if (details[0] instanceof ErrorDetail) {
+        expect(expectedErrorDetail.equals(details[0])).toBeTrue();
       }
     }
   }
@@ -57,30 +56,32 @@ describe("fail_server_streaming", () => {
   const request = new StreamingOutputCallRequest({
     responseParameters: [{ size }],
   });
-  describeTransports(crosstestTransports, (transport, transportName) => {
+  servers.describeTransports((transport) => {
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
       try {
         for await (const response of client.failStreamingOutputCall(request)) {
           expectResponseSize(response);
         }
         fail("expected to catch an error");
       } catch (e) {
-        expectError(e, transportName);
+        expectError(e);
       }
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
       client.failStreamingOutputCall(
         request,
         (response) => {
           expectResponseSize(response);
         },
         (err: ConnectError | undefined) => {
-          expectError(err, transportName);
+          expectError(err);
           done();
         }
       );
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
