+++ packages/connect-node-test/src/crosstest/ping_pong.spec.ts
@@ -14,41 +14,48 @@
 
 import { createPromiseClient } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import { PayloadType } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
+import { interop } from "../helpers/interop.js";
 
-describe("ping_pong", function () {
-  describeTransports(crosstestTransports, (transport) => {
-    const sizes = [31415, 9, 2653, 58979];
-    it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
-      const call = await client.fullDuplexCall();
-      for (const size of sizes) {
-        await call.send({
-          payload: {
-            body: new Uint8Array(size),
-            type: PayloadType.COMPRESSABLE,
-          },
-          responseParameters: [
-            {
-              size: size,
-            },
-          ],
-        });
-        const res = await call.receive();
-        expect(res).toBeDefined();
-        expect(res?.payload).toBeDefined();
-        expect(res?.payload?.type).toBe(PayloadType.COMPRESSABLE);
-        expect(res?.payload?.body.byteLength).toBe(size);
-      }
-      call.close();
-      const res = await call.receive();
-      expect(res).toBeNull();
-    });
-    it("with callback client", function (done) {
-      // TODO(TCN-679) need callback clients
-      done();
-    });
-  });
+describe("ping_pong", () => {
+  const sizes = [31415, 9, 2653, 58979];
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransportsExcluding(
+    [
+      // This test does full duplex, receiving before it closes the input.
+      // In this case, your connect fetch transport never issues the request.
+      "connect fetch transport (binary) against Node.js (http)",
+      "connect fetch transport (JSON) against Node.js (http)",
+    ],
+    (transport) => {
+      it("with promise client", async function () {
+        const client = createPromiseClient(TestService, transport());
+        const stream = await client.fullDuplexCall();
+        for (const size of sizes) {
+          await stream.send({
+            payload: interop.makeServerPayload(PayloadType.COMPRESSABLE, size),
+            responseParameters: [
+              {
+                size,
+                intervalUs: 0,
+              },
+            ],
+          });
+          const res = await stream.receive();
+          expect(res?.payload?.body.length).toBe(size);
+          expect(res?.payload?.type).toBe(PayloadType.COMPRESSABLE);
+        }
+        stream.close();
+      });
+      it("with callback client", function (done) {
+        // TODO(TCN-679, TCN-568) need callback clients
+        done();
+      });
+    }
+  );
+
+  afterAll(async () => await servers.stop());
 });
