+++ packages/connect-node-test/src/crosstest/server_streaming.spec.ts
@@ -17,12 +17,14 @@ import {
   createPromiseClient,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import { StreamingOutputCallRequest } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("server_streaming", function () {
-  describeTransports(crosstestTransports, (transport) => {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransports((transport) => {
     const sizes = [31415, 9, 2653, 58979];
     const request = new StreamingOutputCallRequest({
       responseParameters: sizes.map((size, index) => ({
@@ -31,7 +33,7 @@ describe("server_streaming", function () {
       })),
     });
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
       let responseCount = 0;
       for await (const response of client.streamingOutputCall(request)) {
         expect(response.payload).toBeDefined();
@@ -41,7 +43,7 @@ describe("server_streaming", function () {
       expect(responseCount).toEqual(sizes.length);
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
       let responseCount = 0;
       client.streamingOutputCall(
         request,
@@ -58,4 +60,6 @@ describe("server_streaming", function () {
       );
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
