+++ packages/connect-node-test/src/crosstest/status_code_and_message.spec.ts
@@ -13,18 +13,20 @@
 // limitations under the License.
 
 import {
+  Code,
   ConnectError,
   createCallbackClient,
   createPromiseClient,
-  Code,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import { SimpleRequest } from "../gen/grpc/testing/messages_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("status_code_and_message", function () {
-  describeTransports(crosstestTransports, (transport) => {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransports((transport) => {
     const TEST_STATUS_MESSAGE = "test status message";
     const request = new SimpleRequest({
       responseStatus: {
@@ -40,7 +42,7 @@ describe("status_code_and_message", function () {
       }
     }
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
       try {
         await client.unaryCall(request);
         fail("expected to catch an error");
@@ -49,11 +51,13 @@ describe("status_code_and_message", function () {
       }
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
       client.unaryCall(request, (err: ConnectError | undefined) => {
         expectError(err);
         done();
       });
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
