+++ packages/connect-node-test/src/crosstest/unimplemented_server_streaming_method.spec.ts
@@ -13,27 +13,24 @@
 // limitations under the License.
 
 import {
+  Code,
   ConnectError,
+  connectErrorFromReason,
   createCallbackClient,
   createPromiseClient,
-  Code,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
 import { Empty } from "../gen/grpc/testing/empty_pb.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("unimplemented_server_streaming_method", function () {
-  function expectError(err: unknown) {
-    expect(err).toBeInstanceOf(ConnectError);
-    if (err instanceof ConnectError) {
-      expect(err.code).toEqual(Code.Unimplemented);
-    }
-  }
-  const request = new Empty();
-  describeTransports(crosstestTransports, (transport) => {
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
+
+  servers.describeTransports((transport) => {
     it("with promise client", async function () {
-      const client = createPromiseClient(TestService, transport);
+      const client = createPromiseClient(TestService, transport());
+      const request = new Empty();
       try {
         for await (const response of client.unimplementedStreamingOutputCall(
           request
@@ -42,21 +39,31 @@ describe("unimplemented_server_streaming_method", function () {
         }
         fail("expected to catch an error");
       } catch (e) {
-        expectError(e);
+        expect(e).toBeInstanceOf(ConnectError);
+        expect(connectErrorFromReason(e).code).toBe(Code.Unimplemented);
+        expect(connectErrorFromReason(e).message).toBe(
+          "[unimplemented] grpc.testing.TestService.UnimplementedStreamingOutputCall is not implemented"
+        );
       }
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(TestService, transport);
+      const client = createCallbackClient(TestService, transport());
+      const request = new Empty();
       client.unimplementedStreamingOutputCall(
         request,
         (response) => {
           fail(`expecting no response, got: ${response.toJsonString()}`);
         },
         (err: ConnectError | undefined) => {
-          expectError(err);
+          expect(err?.code).toBe(Code.Unimplemented);
+          expect(err?.message).toBe(
+            "[unimplemented] grpc.testing.TestService.UnimplementedStreamingOutputCall is not implemented"
+          );
           done();
         }
       );
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
