+++ packages/connect-node-test/src/crosstest/unimplemented_service.spec.ts
@@ -13,42 +13,38 @@
 // limitations under the License.
 
 import {
+  Code,
   ConnectError,
+  connectErrorFromReason,
   createCallbackClient,
   createPromiseClient,
-  Code,
 } from "@bufbuild/connect-node";
 import { UnimplementedService } from "../gen/grpc/testing/test_connectweb.js";
-import { describeTransports } from "../helpers/describe-transports.js";
-import { crosstestTransports } from "../helpers/crosstestserver.js";
+import { createTestServers } from "../helpers/testserver.js";
 
 describe("unimplemented_service", function () {
-  function expectError(err: unknown) {
-    // We expect this to be DEADLINE_EXCEEDED, however envoy is monitoring the stream timeout
-    // and will return an HTTP status code 408 when stream max duration time reached, which
-    // cannot be translated to a connect error code.
-    expect(err).toBeInstanceOf(ConnectError);
-    if (err instanceof ConnectError) {
-      expect(err.code === Code.DeadlineExceeded);
-    }
-  }
+  const servers = createTestServers();
+  beforeAll(async () => await servers.start());
 
-  describeTransports(crosstestTransports, (transport) => {
+  servers.describeTransports((transport) => {
     it("with promise client", async function () {
-      const client = createPromiseClient(UnimplementedService, transport);
+      const client = createPromiseClient(UnimplementedService, transport());
       try {
         await client.unimplementedCall({});
         fail("expected to catch an error");
       } catch (e) {
-        expectError(e);
+        expect(e).toBeInstanceOf(ConnectError);
+        expect(connectErrorFromReason(e).code).toBe(Code.Unimplemented);
       }
     });
     it("with callback client", function (done) {
-      const client = createCallbackClient(UnimplementedService, transport);
+      const client = createCallbackClient(UnimplementedService, transport());
       client.unimplementedCall({}, (err: ConnectError | undefined) => {
-        expectError(err);
+        expect(err?.code).toBe(Code.Unimplemented);
         done();
       });
     });
   });
+
+  afterAll(async () => await servers.stop());
 });
