+++ packages/connect-node-test/src/helpers/testserver.ts
@@ -2,157 +2,13 @@ import { createConnectTransport } from "@bufbuild/connect-web-next";
 import * as http2 from "http2";
 import * as http from "http";
 import {
-  Code,
-  ConnectError,
   createConnectHttp2Transport,
   createHandlers,
-  decodeBinaryHeader,
-  encodeBinaryHeader,
   mergeHandlers,
   Transport,
 } from "@bufbuild/connect-node";
 import { TestService } from "../gen/grpc/testing/test_connectweb.js";
-import { interop } from "./interop.js";
-
-/* eslint-disable @typescript-eslint/require-await,require-yield */
-
-const handlers = createHandlers(
-  TestService,
-  {
-    emptyCall() {
-      return {};
-    },
-
-    unaryCall(request, context) {
-      if (request.responseStatus?.code !== undefined) {
-        throw new ConnectError(
-          request.responseStatus.message,
-          request.responseStatus.code
-        );
-      }
-      echoMetadata(
-        context.requestHeader,
-        context.responseHeader,
-        context.responseTrailer
-      );
-      return {
-        payload: interop.makeServerPayload(
-          request.responseType,
-          request.responseSize
-        ),
-      };
-    },
-
-    failUnaryCall() {
-      throw new ConnectError(
-        interop.nonASCIIErrMsg,
-        Code.ResourceExhausted,
-        {},
-        [interop.errorDetail]
-      );
-    },
-
-    cacheableUnaryCall(/*request*/) {
-      throw new ConnectError("TODO", Code.Unimplemented);
-    },
-
-    async *streamingOutputCall(request, context) {
-      echoMetadata(
-        context.requestHeader,
-        context.responseHeader,
-        context.responseTrailer
-      );
-      for (const param of request.responseParameters) {
-        if (param.intervalUs > 0) {
-          await new Promise<void>((resolve) => {
-            setTimeout(resolve, param.intervalUs / 1000);
-          });
-        }
-        yield {
-          payload: interop.makeServerPayload(request.responseType, param.size),
-        };
-      }
-    },
-
-    async *failStreamingOutputCall() {
-      throw new ConnectError(
-        interop.nonASCIIErrMsg,
-        Code.ResourceExhausted,
-        {},
-        [interop.errorDetail]
-      );
-    },
-
-    async streamingInputCall(requests) {
-      let total = 0;
-      for await (const req of requests) {
-        total += req.payload?.body.length ?? 0;
-      }
-      return {
-        aggregatedPayloadSize: total,
-      };
-    },
-
-    async *fullDuplexCall(requests, context) {
-      echoMetadata(
-        context.requestHeader,
-        context.responseHeader,
-        context.responseTrailer
-      );
-      for await (const req of requests) {
-        if (req.responseStatus?.code !== undefined) {
-          throw new ConnectError(
-            req.responseStatus.message,
-            req.responseStatus.code
-          );
-        }
-        for (const param of req.responseParameters) {
-          if (param.intervalUs > 0) {
-            await new Promise<void>((resolve) => {
-              setTimeout(resolve, param.intervalUs / 1000);
-            });
-          }
-          yield {
-            payload: interop.makeServerPayload(req.responseType, param.size),
-          };
-        }
-      }
-    },
-
-    async *halfDuplexCall(/*requests*/) {
-      yield {};
-      throw new ConnectError("TODO", Code.Unimplemented);
-    },
-
-    async *unimplementedCall(/*request*/) {
-      throw new ConnectError("Not implemented", Code.Unimplemented);
-    },
-
-    async *unimplementedStreamingOutputCall(/*requests*/) {
-      throw new ConnectError("Not implemented", Code.Unimplemented);
-    },
-  },
-  {}
-);
-
-function echoMetadata(
-  requestHeader: Headers,
-  responseHeader: Headers,
-  responseTrailer: Headers
-): void {
-  const leadingMetadata = requestHeader.get(interop.leadingMetadataKey);
-  if (leadingMetadata !== null) {
-    responseHeader.set(interop.leadingMetadataKey, leadingMetadata);
-  }
-  const trailingMetadata = requestHeader.get(interop.trailingMetadataKey);
-  if (trailingMetadata !== null) {
-    const decodedTrailingMetadata = decodeBinaryHeader(trailingMetadata);
-    responseTrailer.set(
-      interop.trailingMetadataKey,
-      encodeBinaryHeader(decodedTrailingMetadata)
-    );
-  }
-}
+import { testService } from "./test-service-impl.js";
 
 export function createTestServers() {
   // TODO http2 server with TLS and allow http1
@@ -176,6 +32,32 @@ export function createTestServers() {
   // Node's undici fetch does not support HTTP1.1 (as of Nov 2022), so we cannot
   // run it against the H2C server
   const transports = {
+    // TODO add gRPC-web transport once implemented
+    // TODO add gRPC transport once implemented
+    // TODO add http1.1 transports once implemented
+    // The following servers are available through crosstests:
+    //
+    // | server        | port |
+    // | ------------- | --- |
+    // | connect-go h1 | 8080 |
+    // | connect-go h2 | 8081 |
+    // | grpc-go       | 8083 |
+    //
+    // Source: // https://github.com/bufbuild/connect-web/pull/87
+    "connect Node.js http2 transport (binary) against crosstest (connect-go h1)":
+      (options?: Record<string, unknown>) =>
+        createConnectHttp2Transport({
+          ...options,
+          baseUrl: `https://127.0.0.1:8080`,
+          useBinaryFormat: true,
+        }),
+    "connect Node.js http2 transport (JSON) against crosstest (connect-go h1)":
+      (options?: Record<string, unknown>) =>
+        createConnectHttp2Transport({
+          ...options,
+          baseUrl: `https://127.0.0.1:8080`,
+          useBinaryFormat: false,
+        }),
     "connect Node.js http2 transport (binary) against Node.js (H2C)": (
       options?: Record<string, unknown>
     ) =>
@@ -244,12 +126,18 @@ export function createTestServers() {
       return Promise.all([
         new Promise<void>((resolve) => {
           nodeH2cServer = http2
-            .createServer({}, mergeHandlers(handlers))
+            .createServer(
+              {},
+              mergeHandlers(createHandlers(TestService, testService, {}))
+            )
             .listen(0, resolve);
         }),
         new Promise<void>((resolve) => {
           nodeHttpServer = http
-            .createServer({}, mergeHandlers(handlers))
+            .createServer(
+              {},
+              mergeHandlers(createHandlers(TestService, testService, {}))
+            )
             .listen(0, resolve);
         }),
       ]).then();
